// src/app/themes-custom/interfaces/interface.ts
export interface IProdultListPageData {
  productList: any[];
  currentCategoryName?: string;
  searchProducts: boolean;
  cardInfo: any;
  paginating: boolean;
  hasImages: number;
  layout_type: number;
  isRestaurantActive: any;
}